feat: add front-end (#2)
* add front-end app

* add new dependencies

* rekt

* remove .next

* make Anchor wallet work

* test on localnot

* add landing page

* wip

* add footer

* good enough v1

* fix logout

* wip

* refactor

* add static trades page

* enough for now

* wip

* wip

* add calls

* wip

* wip

* add packages missing

* wip

* add project serum anchor dependency

* wip

* fetch all trades

* wip

* wip

* wip

* wip

* wip

* working createMangoAccount

* wip: deposit USDC to Trade account

* cleanup

* cleanup

+++ programs/tanto/src/state/user.rs
@@ -2,18 +2,18 @@ use anchor_lang::prelude::*;
 
 #[account]
 pub struct User {
-  id: usize,
+  id: i64,
   owner: Pubkey,
   twitter: String,
   avatar: String,
   success_percentage: u8,
-  started_trades: Vec<usize>,
-  funded_trades: Vec<usize>,
+  started_trades: Vec<i64>,
+  funded_trades: Vec<i64>,
   bump: u8,
 }
 
 impl User {
-  pub fn create_user(&mut self, id: usize, twitter: String, bump: u8) -> Result<()> {
+  pub fn create_user(&mut self, id: i64, twitter: String, bump: u8) -> Result<()> {
     self.twitter = twitter;
     self.bump = bump;
